Refactor application for Floats
Calculators can express Floats, so the data model needed to change to
manage state via Floats.
This change complicated the pattern matching in the `update` function.
Elm forbids you from pattern matching on Floats. This is reasonable, but
annoying. In a few places I want to pattern match on `0` which was
impossible, to work around this I changed the case statement to match
against a `truncate`d value. This seems bad and is highlighting to me
that the data model needs refactoring.