Allow user to repeat last operation
Calculators will repeat the previous operation if you press the equals
sign again after evaluation. I had abandoned this feature during initial
implementation b/c there was no way to infer that the application was in
that a state where that action was permissible from the data model.
This expansion of the model lets me infer that state, however it feels
weird and incorrect, and it also opens up opportunities for incorrect or impossible
states. e.g. If the model includes a previous operation, but the user
has created a new one, the new operation will be superseded by the
previous one upon pressing Equals.
This complexity & unreliability seems to be the nail in the coffin for
this data model. Time to refactor.