Add decimal point button
Now you can work with decimal points! This was cake to implement once
the app state was managed via strings.

My primary difficulty with this feature when the app state was managed
via floats was representing the state of having just pressed the decimal
point button. e.g. rendering "1."

The other complexity was being able to detect when the app was in the
state that pressing an operand should append past the decimal point.
Strings made this trivial.