Refactor Input type for simplicity
Limits the variants of the Input type to 4 options to reduce the
complexitiy of state changes.
This introduced the idea of tracking the current input as either the
operand, the operator, or the whole expression.
Depending on the input state the calculator tries to render the most
useful information, either the working operand or the eagerly evaluated
result of the in progress, or submitted operation.