Add AllClear Button
This isn't as ideal was I wanted it. Instead of having a button that
either clears the working operation or the entire state, I only have a
button the clears the entire state. I'm making this concession to keep
things simple, as it's plain that this has gotten a bit out of control